ShellPkg: remove memory leak in file handle list
The shell was not freeing sufficient memory when freeing a list of files.  The structure contained a pointer which was being left behind.  I made a new function to replace the shared freeing function which frees the “Path” member of the SHELL_COMMAND_FILE_HANDLE structure.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Jaben Carsey <jaben.carsey@intel.com>
Reviewed-by: Ruiyu Ni <ruiyu.ni@intel.com>

git-svn-id: svn://svn.code.sf.net/p/edk2/code/trunk/edk2/ShellPkg@15073 6f19259b-4bc3-4df7-8a09-765794883524